requirements.txt
kivy
Pillow
ffpyplayer
moviepy
pyinstaller